+++ packages/design/tailwind/CHANGELOG.md
@@ -15,6 +15,21 @@ This file was generated using [@jscutlery/semver](https://github.com/jscutlery/s
 
 
 
+## [0.21.0](https://github.com/ogcio/govie-ds/compare/tailwind-v0.20.0...tailwind-v0.21.0) (2024-11-25)
+
+
+### Features
+
+* html pagination component ([#332](https://github.com/ogcio/govie-ds/issues/332)) ([aae921f](https://github.com/ogcio/govie-ds/commit/aae921ff7bb799be6bfc75acaf7664272f755841))
+* html toast component ([#335](https://github.com/ogcio/govie-ds/issues/335)) ([0afc349](https://github.com/ogcio/govie-ds/commit/0afc349177d051e799d3b06dc56176419256b16f))
+
+## [0.20.0](https://github.com/ogcio/govie-ds/compare/tailwind-v0.19.0...tailwind-v0.20.0) (2024-11-20)
+
+
+### Features
+
+* toast component ([#331](https://github.com/ogcio/govie-ds/issues/331)) ([1ca741e](https://github.com/ogcio/govie-ds/commit/1ca741e609763ba847f24cfd259a8b727d87a498))
+
 ## [0.19.0](https://github.com/ogcio/govie-ds/compare/tailwind-v0.18.0...tailwind-v0.19.0) (2024-11-18)